requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
chromadb
sentence-transformers
